Rounding

+++ p6/src/main.rs
@@ -11,18 +11,17 @@ fn main() {
     let mut res: Vec<i32> = Vec::new();
 
     for _ in 0..n {
-        let v: Vec<i32> = read_input()
+        let v: Vec<f32> = read_input()
             .split_whitespace()
-            .map(|x| x.parse::<i32>().unwrap())
+            .map(|x| x.parse::<f32>().unwrap())
             .collect();
 
-        let temp: f32 = v[0] as f32 / v[1] as f32;
+        let temp: f32 = v[0] / v[1];
 
         res.push(temp.round() as i32);
     }
 
     //print result
-
     for i in res {
         print!("{} ", i);
     }
